requirements.txt
streamlit>=1.27.0
langchain>=0.0.27
langchain-community>=0.0.10
langchain-openai>=0.0.2
langchain-chroma>=0.0.10
langchain-huggingface>=0.0.2
chromadb>=0.4.18
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
openai>=1.3.0
pymupdf>=1.22.0
tqdm>=4.65.0

# Advanced NLP libraries (optional but recommended)
spacy>=3.7.2
nltk>=3.8.1
bertopic>=0.15.0
scikit-learn>=1.3.2

# Required for LLM-based metadata extraction
pydantic>=2.4.2

# Required for enhanced embedding
langchain-core>=0.1.0 